+++ app/controllers/public_recipes_controller.rb
@@ -1,0 +1,5 @@
+class PublicRecipesController < ApplicationController
+  def index
+    @public_recipes = Recipe.where(public: true).order('created_at DESC')
+  end
+end
