+++ app/helpers/user_helper.rb
@@ -1,0 +1,2 @@
+module UserHelper
+end
