+++ app/views/layouts/_navbar.html.erb
@@ -1,0 +1,23 @@
+<nav class="navbar navbar-expand-lg bg-light">
+  <div class="container-fluid">
+    <span class="navbar-brand mb-0 h1">RecipeOnRails</span>
+    <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+      <span class="navbar-toggler-icon"></span>
+    </button>
+    <div class="collapse navbar-collapse" id="navbarSupportedContent">
+      <ul class="navbar-nav me-auto mb-2 mb-lg-0">
+        <%= link_to "Public Recipes", recipes_index_path, :class => 'nav-link' %>
+        <%= link_to "Users", users_path, :class => 'nav-link' %>
+        <% if user_signed_in? %>
+          <%= link_to "My Recipe List", user_recipes_path(user_id: current_user.id), :class => 'nav-link' %>
+          <%= link_to "My Food List", user_foods_path(user_id: current_user.id), :class => 'nav-link' %>
+          <%= link_to 'Edit profile', edit_user_registration_path, :class => 'nav-link' %>
+          <%= button_to "Logout", destroy_user_session_path, method: :delete, data: { turbo: false }, :class => 'nav-link' %>
+        <% else %>
+          <%= link_to "Sign up", new_user_registration_path, :class => 'nav-link' %>
+          <%= link_to "Log in", new_user_session_path, :class => 'nav-link' %>
+        <% end %>
+      </ul>
+    </div>
+  </div>
+</nav>
