+++ app/views/recipe_food/_update.html.erb
@@ -1,10 +1,10 @@
-<%= form_with model: @recipe_food, url: user_recipe_recipe_food_path do |form| %>
-  <% if @recipe_food.errors.any? %>
+<%= form_with model: @recipe_foods, url: user_recipe_recipe_foods_path do |form| %>
+  <% if @recipe_foods.errors.any? %>
     <div style="color: red">
-      <h2><%= pluralize(@recipe_food.errors.count, "error") %> prohibited this recipe_food from being saved:</h2>
+      <h2><%= pluralize(@recipe_foods.errors.count, "error") %> prohibited this recipe_foods from being saved:</h2>
 
       <ul>
-        <% @recipe_food.errors.each do |error| %>
+        <% @recipe_foods.errors.each do |error| %>
           <li><%= error.full_message %></li>
         <% end %>
       </ul>
