+++ app/views/recipe_food/edit.html.erb
@@ -1,6 +1,6 @@
 <div class="recipeFoodEditContainer">
 <h1 class="recipeFoodEdit">Edit Recipe Food</h1>
-<%= render "update", recipe_food: @recipe_food %>
+<%= render "update", recipe_foods: @recipe_foods %>
 <br>
 <div>
   <%= link_to "Back to recipe foods", user_recipe_path(@user, @recipe) %>
