+++ app/views/recipe_food/new.html.erb
@@ -1,5 +1,5 @@
 <div class="newRecipeFood">
 <h1 class="newRecipeFoodTitle">Add a new food to your recipe</h1>
 
-<%= render "form", recipe_food: @recipe_food %>
+<%= render "form", recipe_foods: @recipe_foods %>
 </div>
